+++ custom/icds_reports/utils/aggregation_helpers/monolith/agg_awc.py
@@ -11,7 +11,7 @@ from corehq.util.python_compatibility import soft_assert_type_text
 from custom.icds_reports.const import AGG_CCS_RECORD_CF_TABLE
 from custom.icds_reports.utils.aggregation_helpers import transform_day_to_month, \
     month_formatter
-from custom.icds_reports.utils.aggregation_helpers.monolith import BaseICDSAggregationHelper
+from custom.icds_reports.utils.aggregation_helpers.monolith.base import BaseICDSAggregationHelper
 
 
 class AggAwcHelper(BaseICDSAggregationHelper):
