+++ custom/icds_reports/utils/aggregation_helpers/monolith/awc_location.py
@@ -6,7 +6,7 @@ from six.moves import map
 from corehq.apps.userreports.models import StaticDataSourceConfiguration, get_datasource_config
 from corehq.apps.userreports.util import get_table_name
 from custom.icds_reports.const import AWC_LOCATION_TABLE_ID, AWW_USER_TABLE_ID
-from custom.icds_reports.utils.aggregation_helpers.monolith import BaseICDSAggregationHelper
+from custom.icds_reports.utils.aggregation_helpers.monolith.base import BaseICDSAggregationHelper
 
 
 class LocationAggregationHelper(BaseICDSAggregationHelper):
