+++ custom/icds_reports/utils/aggregation_helpers/monolith/child_health_monthly.py
@@ -14,7 +14,7 @@ from custom.icds_reports.const import (
 )
 from custom.icds_reports.utils.aggregation_helpers import transform_day_to_month, \
     month_formatter
-from custom.icds_reports.utils.aggregation_helpers.monolith import BaseICDSAggregationHelper
+from custom.icds_reports.utils.aggregation_helpers.monolith.base import BaseICDSAggregationHelper
 
 
 class ChildHealthMonthlyAggregationHelper(BaseICDSAggregationHelper):
