+++ custom/icds_reports/utils/aggregation_helpers/monolith/daily_feeding_forms_child_health.py
@@ -5,7 +5,7 @@ from dateutil.relativedelta import relativedelta
 
 from custom.icds_reports.const import AGG_DAILY_FEEDING_TABLE
 from custom.icds_reports.utils.aggregation_helpers import month_formatter
-from custom.icds_reports.utils.aggregation_helpers.monolith import BaseICDSAggregationHelper
+from custom.icds_reports.utils.aggregation_helpers.monolith.base import BaseICDSAggregationHelper
 
 
 class DailyFeedingFormsChildHealthAggregationHelper(BaseICDSAggregationHelper):
